chore: add Lingui configuration file for localizations
A new file, lingui.config.js, was added to handle localization. This configures lingui to support multiple locales (ar, en, fa, tr) with English as the source. Messages for each locale are stored under '<rootDir>/src/locales/{locale}/messages' and node_modules, .next, and out directories are excluded from localization. This setup is intended to facilitate the addition of new locales and the translation of messages.