refactor: add utility files to handle locale and style merge
Introduced new utility files to handle locale detection and merging of class styles. In 'locales.ts', created methods to get, load and work with different locales based on the lingui configurations. In 'index.ts', a new method was added to merge styles using clsx and tailwind-merge, improving the handling of dynamic classnames in components.