ci: update GitHub actions repository condition
Changed the Github actions to now run only when the repository name is 'lilnouns/lilnouns-template' instead of 'nekofar/nextjs-wagmi-template'. This is to reflect the shift in the repository we'll be using for our CI/CD process. This change ensures that only valid commits to the updated repository will trigger the actions.